pyproject.toml
[project]
name = "local-search-mcp"
version = "0.1.0"
description = "Offline Wikipedia Search MCP Server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "rank-bm25>=0.2.2",
    "datasets>=2.14.0",
    "tqdm>=4.66.0",
    "ollama>=0.1.0",
    "numpy>=1.24.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "langchain-text-splitters>=1.1.0",
    "langchain-core>=1.2.6",
    "langdetect>=1.0.9",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
